requirements.txt
click==6.6
configparser==3.5.0
cssmin==0.2.0
Fabric3==1.14.post1
Flask==1.0.2
Flask-Assets==0.12
Flask-Babel==0.11.1
-e git+https://github.com/nickwynja/Flask-FlatPages#egg=Flask_FlatPages
Flask-Imagine==0.5
Flask-Markdown==0.3
-e git+https://github.com/nickwynja/Frozen-Flask#egg=Frozen-Flask
Jinja2==2.10
jsmin==2.2.2
libsass==0.15.1
livereload==2.5.1
Markdown==3.0.1
MarkupSafe==0.23
mdx_poetic
Pygments==2.1.3
pymdown-extensions==6.0.0
python-dotenv==0.9.1
python-frontmatter==0.4.2
python-markdown-comments==0.0.0
pytz==2016.7
PyYAML==3.13
webassets==0.12.1
webassets-libsass==0.2
